Consolidate the pre-exec paths between simulation and inference
There are a number of transformations that are applied to a
ThingTalk program between the user's issuing it, and it being
executed: we assign device IDs, resolve relative locations and
times, lookup contacts by name, etc.
It is important to minimize any difference between simulation
and inference time on these, otherwise we train on bogus
contexts that don't match the contexts we see at inference itime.

Consolidate the code so that simulation uses the real code
paths as much as possible. The code is factored into an
abstract class AbstractDialogueAgent common to both, and
subclasses for simulaton and execution. The abstract class
incorporates all logic that was previously in the the
entity-linking folder (aka the "legacy dialogues" of old
almond-dialog-agent). The actual interaction with the engine
or the user is abstracted out, and simulated when running
in simulation (generating training data).

All of this is so I can say "get weather for san francisco"
when annotating data in "manual-annotate-data", and actually
see the dialogue continue instead of crashing with a compiler
error.